[ML] Data Frames: Improve start/stop behavior (#36757)
- Adds a confirm modal to the Start button of the data frame jobs list. image
- Adds a snippet with a warning about possibly cluster load to wizard. image
- Uses _stop?force=true to stop jobs.
- Adds a check to disable the start button in the jobs list if the job is a completed batch job.
