+++ x-pack/plugins/ml/public/data_frame/components/job_create/job_create_form.tsx
@@ -211,16 +211,19 @@ export const JobCreateForm: SFC<Props> = React.memo(
       return `PUT _data_frame/transforms/${jobId}\n${JSON.stringify(jobConfig, null, 2)}\n\n`;
     }
 
-    const ITEM_STYLE = { width: '300px' };
+    // TODO move this to SASS
+    const FLEX_GROUP_STYLE = { height: '90px', maxWidth: '800px' };
+    const FLEX_ITEM_STYLE = { width: '200px' };
+    const PANEL_ITEM_STYLE = { width: '300px' };
 
     return (
       <EuiForm>
         {!created && (
-          <EuiFlexGroup alignItems="center" style={{ maxWidth: '800px' }}>
-            <EuiFlexItem grow={false} style={{ width: '200px' }}>
+          <EuiFlexGroup alignItems="center" style={FLEX_GROUP_STYLE}>
+            <EuiFlexItem grow={false} style={FLEX_ITEM_STYLE}>
               <EuiButton fill isDisabled={created && started} onClick={createAndStartDataFrame}>
                 {i18n.translate('xpack.ml.dataframe.jobCreateForm.createAndStartDataFrameButton', {
-                  defaultMessage: 'Create & start',
+                  defaultMessage: 'Create and start',
                 })}
               </EuiButton>
             </EuiFlexItem>
@@ -230,7 +233,7 @@ export const JobCreateForm: SFC<Props> = React.memo(
                   'xpack.ml.dataframe.jobCreateForm.createAndStartDataFrameDescription',
                   {
                     defaultMessage:
-                      'Create and start the data frame job. After the job is started, you will be offered options to continue exploring the data frame job.',
+                      'Creates and starts the data frame job. A data frame job will increase search and indexing load in your cluster. Please stop the job if excessive load is experienced. After the job is started, you will be offered options to continue exploring the data frame job.',
                   }
                 )}
               </EuiText>
@@ -238,9 +241,9 @@ export const JobCreateForm: SFC<Props> = React.memo(
           </EuiFlexGroup>
         )}
         {created && (
-          <EuiFlexGroup alignItems="center" style={{ maxWidth: '800px' }}>
-            <EuiFlexItem grow={false} style={{ width: '200px' }}>
-              <EuiButton isDisabled={created && started} onClick={startDataFrame}>
+          <EuiFlexGroup alignItems="center" style={FLEX_GROUP_STYLE}>
+            <EuiFlexItem grow={false} style={FLEX_ITEM_STYLE}>
+              <EuiButton fill isDisabled={created && started} onClick={startDataFrame}>
                 {i18n.translate('xpack.ml.dataframe.jobCreateForm.startDataFrameButton', {
                   defaultMessage: 'Start',
                 })}
@@ -250,14 +253,14 @@ export const JobCreateForm: SFC<Props> = React.memo(
               <EuiText color="subdued" size="s">
                 {i18n.translate('xpack.ml.dataframe.jobCreateForm.startDataFrameDescription', {
                   defaultMessage:
-                    'Starts the data frame job. After the job is started, you will be offered options to continue exploring the data frame job.',
+                    'Starts the data frame job. A data frame job will increase search and indexing load in your cluster. Please stop the job if excessive load is experienced. After the job is started, you will be offered options to continue exploring the data frame job.',
                 })}
               </EuiText>
             </EuiFlexItem>
           </EuiFlexGroup>
         )}
-        <EuiFlexGroup alignItems="center" style={{ maxWidth: '800px' }}>
-          <EuiFlexItem grow={false} style={{ width: '200px' }}>
+        <EuiFlexGroup alignItems="center" style={FLEX_GROUP_STYLE}>
+          <EuiFlexItem grow={false} style={FLEX_ITEM_STYLE}>
             <EuiButton isDisabled={created} onClick={createDataFrame}>
               {i18n.translate('xpack.ml.dataframe.jobCreateForm.createDataFrameButton', {
                 defaultMessage: 'Create',
@@ -273,8 +276,8 @@ export const JobCreateForm: SFC<Props> = React.memo(
             </EuiText>
           </EuiFlexItem>
         </EuiFlexGroup>
-        <EuiFlexGroup alignItems="center" style={{ maxWidth: '800px' }}>
-          <EuiFlexItem grow={false} style={{ width: '200px' }}>
+        <EuiFlexGroup alignItems="center" style={FLEX_GROUP_STYLE}>
+          <EuiFlexItem grow={false} style={FLEX_ITEM_STYLE}>
             <EuiCopy textToCopy={getJobConfigDevConsoleStatement()}>
               {(copy: () => void) => (
                 <EuiButton onClick={copy} style={{ width: '100%' }}>
@@ -324,7 +327,7 @@ export const JobCreateForm: SFC<Props> = React.memo(
           <Fragment>
             <EuiHorizontalRule />
             <EuiFlexGrid gutterSize="l">
-              <EuiFlexItem style={ITEM_STYLE}>
+              <EuiFlexItem style={PANEL_ITEM_STYLE}>
                 <EuiCard
                   icon={<EuiIcon size="xxl" type="list" />}
                   title={i18n.translate('xpack.ml.dataframe.jobCreateForm.jobsListCardTitle', {
@@ -340,7 +343,7 @@ export const JobCreateForm: SFC<Props> = React.memo(
                 />
               </EuiFlexItem>
               {started === true && createIndexPattern === true && indexPatternId === undefined && (
-                <EuiFlexItem style={ITEM_STYLE}>
+                <EuiFlexItem style={PANEL_ITEM_STYLE}>
                   <EuiPanel style={{ position: 'relative' }}>
                     <EuiProgress size="xs" color="primary" position="absolute" />
                     <EuiText color="subdued" size="s">
@@ -357,7 +360,7 @@ export const JobCreateForm: SFC<Props> = React.memo(
                 </EuiFlexItem>
               )}
               {started === true && indexPatternId !== undefined && (
-                <EuiFlexItem style={ITEM_STYLE}>
+                <EuiFlexItem style={PANEL_ITEM_STYLE}>
                   <EuiCard
                     icon={<EuiIcon size="xxl" type="discoverApp" />}
                     title={i18n.translate('xpack.ml.dataframe.jobCreateForm.discoverCardTitle', {
